Writing to SQL Database working
This should be enough to get it to actually run. Writing to the SQL data base does work with these files. If it doesn't work try to see that you are using the same java version, and make sure the main file is using the correct driver which is jdbc-3.49.10.jar. These are two issues that I encountered, that I had to fix.